+++ src/macroPseudoterminal.ts
@@ -186,6 +186,7 @@ export class MacroPseudoterminal implements vscode.Pseudoterminal {
         if (history?.length) {
           await createMacro(this.context, undefined, {
             content: `\n// History: ${new Date().toLocaleString()}\n\n` + history.join('\n'),
+            language: this.useTs ? 'typescript' : 'javascript',
             preserveFocus: true,
           });
         } else {